+++ src/utils/dimensionUtils.ts
@@ -2,11 +2,9 @@ import { DimensionSchemaV2025 } from "sailpoint-api-client";
 import { convertPascalCase2SpaceBased, isEmpty } from "./stringUtils";
 import { CSV_MULTIVALUE_SEPARATOR } from "../constants";
 
-export function dimensionSchemaToString(dimensionSchema: DimensionSchemaV2025) {
-
+export function dimensionSchemaToString(dimensionSchema: DimensionSchemaV2025 | undefined) {
     return dimensionSchema?.dimensionAttributes?.map(x => x.name).join(CSV_MULTIVALUE_SEPARATOR)
 }
-
 
 export function stringToDimensionAttributes(input: string): DimensionSchemaV2025 | undefined {
     if (isEmpty(input)) {